Add option for empty default values
In Quarkus at static init time we build what is
essentially a no-op CP instance, but creating
this instance currently incurs some overhead
due to the use of Config to look up default values.
This change, will allow Quarkus to set empty
default values option
